+++ src/algorithms/rem_factorial.tpp
@@ -11,7 +11,7 @@ using NTL::ZZ_pX;
 
 
 template <typename T, typename U>
-T calculate_factorial(long n, const U& m, const std::function<vector<T> (long, long)>& get_A, const PolyMatrix& formula = {}) {
+T calculate_factorial(long long n, const U& m, const std::function<vector<T> (long long, long long)>& get_A, const PolyMatrix& formula = {}) {
     return naive_factorial(n, m, get_A); // TODO: replace temporary measure with something better
 }
