+++ src/algorithms/utils.hpp
@@ -12,9 +12,9 @@
  */
 typedef std::vector<std::vector<std::vector<std::vector<bool>>>> PolyMatrix; 
 
-bool is_power2(long);
+bool is_power2(long long);
 
-long flatten(long, long);
+long long flatten(long long, long long);
 
 /*
  * TODO: this returns a smaller vector of just the subtree rooted at that node.
@@ -25,12 +25,12 @@ long flatten(long, long);
  * how it interacts with incomplete trees.
  */
 template <typename T>
-std::vector<T> get_subtree (const std::vector<T>&, long);
+std::vector<T> get_subtree (const std::vector<T>&, long long);
 
 //Similar to above. Returns the slice of the input vector which corresponds to
 //the leaves of a subtree rooted at k.
 template <typename T>
-std::vector<T> get_shade(const std::vector<T>&, long);
+std::vector<T> get_shade(const std::vector<T>&, long long);
 
 template<typename T>
 std::vector<T> elt_to_base(const std::vector<Elt<T>>&);
