+++ src/searches/benchmarks.cpp
@@ -1,4 +1,4 @@
-//std::function< std::vector<bool> (long N)> remember this is what is being passed in!
+//std::function< std::vector<bool> (long long N)> remember this is what is being passed in!
 #include <NTL/ZZ.h>
 
 #include "../elements/Element.hpp"
@@ -9,9 +9,9 @@
 
 using std::vector;
 
-vector<bool> constant_slow(long B) {
-	long n = 1;
-	for (long i = 0; i < 10*B; ++i) { n = (((n*2+17)/3-19)*5-101); } //just gibberish
+vector<bool> constant_slow(long long B) {
+	long long n = 1;
+	for (long long i = 0; i < 10*B; ++i) { n = (((n*2+17)/3-19)*5-101); } //just gibberish
 
 	vector<bool> incidence;
 	bool ans = (B+n)%5;
@@ -21,7 +21,7 @@ vector<bool> constant_slow(long B) {
 }
 
 
-vector<bool> random_zz(long B, long forest_param, long recompute_param) {
+vector<bool> random_zz(long long B, long long forest_param, long long recompute_param) {
 
 	vector <Elt<NTL::ZZ> > A_rand (B);
 	vector <Elt<NTL::ZZ> > m_rand (B);
@@ -30,10 +30,10 @@ vector<bool> random_zz(long B, long forest_param, long recompute_param) {
 
 	std::random_device rd;
 	std::mt19937 mt(rd());
-	std::uniform_int_distribution<long> dist(1, B);
+	std::uniform_int_distribution<long long> dist(1, B);
 
-	for(long i = 0; i < B; ++i){
-		//long bitsize = log2(i+1)+2;
+	for(long long i = 0; i < B; ++i){
+		//long long bitsize = log2(i+1)+2;
 		A_rand[i] = dist(mt);
 		m_rand[i] = dist(mt);
 	}
